+++ include/ignition/transport/Node.hh
@@ -115,6 +115,31 @@ namespace ignition
         /// \return true when success.
         public: bool Publish(const ProtoMsg &_msg);
 
+        /// \brief Publish a raw pre-serialized message.
+        ///
+        /// \warning This function is only intended for advanced users. The
+        /// standard publishing function, Publish(const ProtoMsg &_msg), will
+        /// ensure that your message is correctly serialized. It is strongly
+        /// recommended that you use the standard publishing function unless
+        /// there is a specific reason for using this one (e.g. you are
+        /// forwarding or playing back data instead of serializing/deserializing
+        /// it). We currently only support the serialization scheme of protobuf.
+        ///
+        /// \warning This function will copy the message data when
+        /// publishing to remote subscribers (interprocess communication).
+        ///
+        /// \note This function will deserialize the message when sending it to
+        /// local (intraprocess) subscribers.
+        ///
+        /// \param[in] _msgData A std::string that represents a
+        /// serialized google::protobuf message.
+        /// \param[in] _msgType A std::string that contains the message type
+        /// name.
+        /// \return true when success.
+        public: bool PublishRaw(
+          const std::string &_msgData,
+          const std::string &_msgType);
+
         /// \brief Check if message publication is throttled. If so, verify
         /// whether the next message should be published or not.
         /// \return true if the message should be published or false otherwise.
@@ -127,7 +152,7 @@ namespace ignition
         /// \internal
         /// \brief Smart pointer to private data.
         /// This is std::shared_ptr because we want to trigger the destructor
-        /// only once when all references to PublisherPrivate are out of scope.
+        /// only once: when all references to PublisherPrivate are out of scope.
         /// The destructor of PublisherPrivate unadvertise the topic.
         private: std::shared_ptr<PublisherPrivate> dataPtr;
       };
@@ -315,7 +340,7 @@ namespace ignition
         // associated with a topic. When the receiving thread gets new data,
         // it will recover the subscription handler associated to the topic and
         // will invoke the callback.
-        this->Shared()->localSubscriptions.AddHandler(
+        this->Shared()->localSubscribers.normal.AddHandler(
           fullyQualifiedTopic, this->NodeUuid(), subscrHandlerPtr);
 
         return this->SubscribeHelper(fullyQualifiedTopic);
@@ -539,7 +564,7 @@ namespace ignition
             [_cb](const RequestT &_req, ReplyT &_rep)
         {
           bool internalResult = false;
-          (*_cb)(_req, _rep, internalResult);
+          (_cb)(_req, _rep, internalResult);
           return internalResult;
         };
 
@@ -1215,6 +1240,25 @@ namespace ignition
       /// \return False if unable to get service info.
       public: bool ServiceInfo(const std::string &_service,
                               std::vector<ServicePublisher> &_publishers) const;
+
+      /// \brief Subscribe to a topic registering a callback. The callback must
+      /// accept a std::string to represent the message data, and a MessageInfo
+      /// which provides metadata about the message.
+      /// \param[in] _topic Name of the topic to subscribe to
+      /// \param[in] _callback A function pointer or std::function object that
+      /// has a void return value and accepts two arguments:
+      /// (const std::string &_msgData, const MessageInfo &_info).
+      /// \param[in] _msgType The type of message to subscribe to. Using
+      /// kGenericMessageType (the default) will allow this subscriber to listen
+      /// to all message types. The callback function can identify the type for
+      /// each message by inspecting its const MessageInfo& input argument.
+      /// \param[in] _opts Options for subscribing.
+      /// \return True if subscribing was successful.
+      public: bool SubscribeRaw(
+        const std::string &_topic,
+        const RawCallback &_callback,
+        const std::string &_msgType = kGenericMessageType,
+        const SubscribeOptions &_opts = SubscribeOptions());
 
       /// \brief Get the partition name used by this node.
       /// \return The partition name.