+++ include/ignition/transport/TopicUtils.hh
@@ -58,12 +58,39 @@ namespace ignition
 
       /// \brief Get the full topic path given a namespace and a topic name.
       /// A fully qualified topic name's length must not exceed kMaxNameLength.
+      /// The fully qualified name follows the next syntax:
+      /// @<PARTITION>@<NAMESPACE>/<TOPIC>
+      /// where:
+      /// <PARTITION>: The name of the partition or empty string.
+      ///              A "/" will be prefixed to the partition name unless is
+      ///              empty or it already starts with slash. A trailing slash
+      ///              will always be removed.
+      /// <NAMESPACE>: The namespace or empty string. A namespace is a prefix
+      ///              applied to the topic name. If not empty, it will always
+      ///              start with a "/". A trailing slash will always be
+      ///              removed
+      /// <TOPIC>:     The topic name. A trailing slash will always be removed.
+      ///
+      /// Note: Intuitively, you can imagine the fully qualified name as a
+      /// UNIX absolute path, where the partition is always sorrounded by "@".
+      /// A namespace, if present, corresponds with the directories of the
+      /// path, and you can imagine the topic as the filename.
+      ///
+      /// E.g.:
+      ///   Only topic:                @@/topic
+      ///   No partition:              @@/namespace/topic1
+      ///   No namespace:              @/partition@/topic1
+      ///   Partition+namespace+topic: @/my_partition@/name/space/topic
+      ///
       /// \param[in] _partition Partition name.
       /// \param[in] _ns Namespace.
       /// \param[in] _topic Topic name.
       /// \param[out] _name Fully qualified topic name.
       /// \return True if the fully qualified name is valid
       /// (if partition, namespace and topic are correct).
+      /// \sa IsValidPartition
+      /// \sa IsValidNamespace
+      /// \sa IsValidTopic
       public: static bool FullyQualifiedName(const std::string &_partition,
                                              const std::string &_ns,
                                              const std::string &_topic,