+++ src/NodeShared.cc
@@ -228,21 +228,20 @@ NodeShared::~NodeShared()
 void NodeShared::RunReceptionTask()
 {
   bool exitLoop = false;
-
-  // Poll socket for a reply, with timeout.
-  std::vector<zmq::pollitem_t> items =
-  {
-    {static_cast<void*>(*this->dataPtr->subscriber), 0, ZMQ_POLLIN, 0},
-    {static_cast<void*>(*this->dataPtr->control), 0, ZMQ_POLLIN, 0},
-    {static_cast<void*>(*this->dataPtr->replier), 0, ZMQ_POLLIN, 0},
-    {static_cast<void*>(*this->dataPtr->responseReceiver), 0, ZMQ_POLLIN, 0}
-  };
-
   while (!exitLoop)
   {
+    // Poll socket for a reply, with timeout.
+    zmq::pollitem_t items[] =
+    {
+      {static_cast<void*>(*this->dataPtr->subscriber), 0, ZMQ_POLLIN, 0},
+      {static_cast<void*>(*this->dataPtr->control), 0, ZMQ_POLLIN, 0},
+      {static_cast<void*>(*this->dataPtr->replier), 0, ZMQ_POLLIN, 0},
+      {static_cast<void*>(*this->dataPtr->responseReceiver), 0, ZMQ_POLLIN, 0}
+    };
     try
     {
-      zmq::poll(&items[0], items.size(), NodeSharedPrivate::Timeout);
+      zmq::poll(&items[0], sizeof(items) / sizeof(items[0]),
+                NodeSharedPrivate::Timeout);
     }
     catch(...)
     {
@@ -269,8 +268,11 @@ void NodeShared::RunReceptionTask()
 }
 
 //////////////////////////////////////////////////
-bool NodeShared::Publish(const std::string &_topic, char *_data,
-  const size_t _dataSize, DeallocFunc *_ffn, const std::string &_msgType)
+bool NodeShared::Publish(
+    const std::string &_topic,
+    char *_data,
+    const size_t _dataSize, DeallocFunc *_ffn,
+    const std::string &_msgType)
 {
   try
   {
@@ -305,10 +307,7 @@ void NodeShared::RecvMsgUpdate()
   // std::string sender;
   std::string data;
   std::string msgType;
-  std::map<std::string, ISubscriptionHandler_M> handlers;
-  ISubscriptionHandlerPtr firstSubscriberPtr;
-  bool handlersFound;
-  bool firstHandlerFound;
+  HandlerInfo handlerInfo;
 
   {
     std::lock_guard<std::recursive_mutex> lock(this->mutex);
@@ -338,37 +337,125 @@ void NodeShared::RecvMsgUpdate()
       return;
     }
 
-    handlersFound = this->localSubscriptions.Handlers(topic, handlers);
-    firstHandlerFound = this->localSubscriptions.FirstHandler(topic, msgType,
-      firstSubscriberPtr);
+    handlerInfo = this->CheckHandlerInfo(topic);
   }
 
-  // Execute the callbacks registered.
-  if (handlersFound && firstHandlerFound)
+  this->TriggerSubscriberCallbacks(topic, data, msgType, handlerInfo);
+}
+
+//////////////////////////////////////////////////
+NodeShared::HandlerInfo NodeShared::CheckHandlerInfo(
+    const std::string &_topic) const
+{
+  HandlerInfo info;
+
+  std::lock_guard<std::recursive_mutex> lk(this->mutex);
+
+  info.haveLocal = this->localSubscribers.normal.Handlers(
+        _topic, info.localHandlers);
+
+  info.haveRaw = this->localSubscribers.raw.Handlers(
+        _topic, info.rawHandlers);
+
+  return info;
+}
+
+//////////////////////////////////////////////////
+NodeShared::SubscriberInfo NodeShared::CheckSubscriberInfo(
+    const std::string &_topic,
+    const std::string &_msgType) const
+{
+  SubscriberInfo info;
+
+  std::lock_guard<std::recursive_mutex> lk(this->mutex);
+
+  info.haveLocal = this->localSubscribers.normal.Handlers(
+        _topic, info.localHandlers);
+
+  info.haveRaw = this->localSubscribers.raw.Handlers(
+        _topic, info.rawHandlers);
+
+  info.haveRemote = this->remoteSubscribers.HasTopic(
+        _topic, _msgType);
+
+  return info;
+}
+
+//////////////////////////////////////////////////
+void NodeShared::TriggerSubscriberCallbacks(
+    const std::string &_topic,
+    const std::string &_msgData,
+    const std::string &_msgType,
+    const HandlerInfo &_handlerInfo)
+{
+  if (!_handlerInfo.haveLocal && !_handlerInfo.haveRaw)
+    return;
+
+  MessageInfo info;
+  info.SetTopicAndPartition(_topic);
+  info.SetType(_msgType);
+
+  if (_handlerInfo.haveRaw)
   {
-    // Create the message.
-    auto recvMsg = firstSubscriberPtr->CreateMsg(data, msgType);
-    if (!recvMsg)
-      return;
-
-    // Create and populate the message information object.
-    MessageInfo info;
-    info.SetTopicAndPartition(topic);
-    info.SetType(msgType);
-
-    for (const auto &node : handlers)
+    for (const auto &node : _handlerInfo.rawHandlers)
     {
       for (const auto &handler : node.second)
       {
-        ISubscriptionHandlerPtr subscriptionHandlerPtr = handler.second;
-        if (subscriptionHandlerPtr)
+        const RawSubscriptionHandlerPtr &rawHandler = handler.second;
+        if (rawHandler)
         {
-          if (subscriptionHandlerPtr->TypeName() == msgType ||
-              subscriptionHandlerPtr->TypeName() == kGenericMessageType)
-            subscriptionHandlerPtr->RunLocalCallback(*recvMsg, info);
+          if (rawHandler->TypeName() == _msgType ||
+              rawHandler->TypeName() == kGenericMessageType)
+          {
+            rawHandler->RunRawCallback(_msgData.c_str(), _msgData.size(), info);
+          }
         }
         else
-          std::cerr << "Subscription handler is NULL" << std::endl;
+          std::cerr << "Raw subscription handler is NULL" << std::endl;
+      }
+    }
+  }
+
+  if (_handlerInfo.haveLocal)
+  {
+    // This will be instantiated by the first suitable handler that we
+    // encounter. If there is no suitable handler, then we can avoid
+    // deserializing the message altogether.
+    std::shared_ptr<ProtoMsg> msg;
+
+    for (const auto &node : _handlerInfo.localHandlers)
+    {
+      for (const auto &handler : node.second)
+      {
+        const ISubscriptionHandlerPtr &localHandler = handler.second;
+        if (localHandler)
+        {
+          if (localHandler->TypeName() == _msgType ||
+              localHandler->TypeName() == kGenericMessageType)
+          {
+            if (!msg)
+            {
+              // If the message has not been deserialized yet, do it now since
+              // we have allegedly found a subscriber which should be able to
+              // do it.
+              msg = localHandler->CreateMsg(_msgData, _msgType);
+
+              if (!msg)
+              {
+                // If the message could not be created, then none of the
+                // handlers in this process will be able to create it, because
+                // protobuf has access to all message types that the current
+                // process is linked to. If CreateMsg(~,~) fails, then we may
+                // as well quit.
+                return;
+              }
+            }
+
+            localHandler->RunLocalCallback(*msg, info);
+          }
+        }
+        else
+          std::cerr << "Local subscription handler is NULL" << std::endl;
       }
     }
   }
@@ -834,7 +921,7 @@ void NodeShared::OnNewConnection(const MessagePublisher &_pub)
   }
 
   // Check if we are interested in this topic.
-  if (this->localSubscriptions.HasHandlersForTopic(topic) &&
+  if (this->localSubscribers.HasSubscriber(topic) &&
       this->pUuid.compare(procUuid) != 0)
   {
     try
@@ -869,44 +956,32 @@ void NodeShared::OnNewConnection(const MessagePublisher &_pub)
 
       std::this_thread::sleep_for(std::chrono::milliseconds(100));
 
-      std::map<std::string, ISubscriptionHandler_M> handlers;
-      if (this->localSubscriptions.Handlers(topic, handlers))
+      std::vector<std::string> handlerNodeUuids =
+          this->localSubscribers.NodeUuids(topic, _pub.MsgTypeName());
+
+      for (const std::string &nodeUuid : handlerNodeUuids)
       {
-        for (auto const &node : handlers)
-        {
-          for (auto const &handler : node.second)
-          {
-            if (handler.second->TypeName() != kGenericMessageType &&
-                handler.second->TypeName() != _pub.MsgTypeName())
-            {
-              continue;
-            }
+        zmq::message_t msg;
+        msg.rebuild(topic.size());
+        memcpy(msg.data(), topic.data(), topic.size());
+        socket.send(msg, ZMQ_SNDMORE);
 
-            std::string nodeUuid = handler.second->NodeUuid();
+        msg.rebuild(this->pUuid.size());
+        memcpy(msg.data(), this->pUuid.data(), this->pUuid.size());
+        socket.send(msg, ZMQ_SNDMORE);
 
-            zmq::message_t msg;
-            msg.rebuild(topic.size());
-            memcpy(msg.data(), topic.data(), topic.size());
-            socket.send(msg, ZMQ_SNDMORE);
+        msg.rebuild(nodeUuid.size());
+        memcpy(msg.data(), nodeUuid.data(), nodeUuid.size());
+        socket.send(msg, ZMQ_SNDMORE);
 
-            msg.rebuild(this->pUuid.size());
-            memcpy(msg.data(), this->pUuid.data(), this->pUuid.size());
-            socket.send(msg, ZMQ_SNDMORE);
+        msg.rebuild(type.size());
+        memcpy(msg.data(), type.data(), type.size());
+        socket.send(msg, ZMQ_SNDMORE);
 
-            msg.rebuild(nodeUuid.size());
-            memcpy(msg.data(), nodeUuid.data(), nodeUuid.size());
-            socket.send(msg, ZMQ_SNDMORE);
-
-            msg.rebuild(type.size());
-            memcpy(msg.data(), type.data(), type.size());
-            socket.send(msg, ZMQ_SNDMORE);
-
-            std::string data = std::to_string(NewConnection);
-            msg.rebuild(data.size());
-            memcpy(msg.data(), data.data(), data.size());
-            socket.send(msg, 0);
-          }
-        }
+        std::string data = std::to_string(NewConnection);
+        msg.rebuild(data.size());
+        memcpy(msg.data(), data.data(), data.size());
+        socket.send(msg, 0);
       }
     }
     // The remote node might not be available when we are connecting.
@@ -1101,6 +1176,79 @@ bool NodeShared::AdvertisePublisher(const ServicePublisher &_publisher)
 }
 
 //////////////////////////////////////////////////
+bool NodeShared::HandlerWrapper::HasSubscriber(
+    const std::string &_fullyQualifiedTopic,
+    const std::string &_msgType) const
+{
+  std::shared_ptr<ISubscriptionHandler> normalSubscriberPtr;
+  std::shared_ptr<RawSubscriptionHandler> rawSubscriberPtr;
+
+  return this->normal.FirstHandler(
+            _fullyQualifiedTopic, _msgType, normalSubscriberPtr)
+         || this->raw.FirstHandler(
+            _fullyQualifiedTopic, _msgType, rawSubscriberPtr);
+}
+
+//////////////////////////////////////////////////
+bool NodeShared::HandlerWrapper::HasSubscriber(
+    const std::string &_fullyQualifiedTopic) const
+{
+  return this->normal.HasHandlersForTopic(_fullyQualifiedTopic)
+      || this->raw.HasHandlersForTopic(_fullyQualifiedTopic);
+}
+
+//////////////////////////////////////////////////
+template <typename HandlerT>
+static void AppendNodeUuids(const HandlerStorage<HandlerT> &_handlerStorage,
+                            const std::string &_fullyQualifiedTopic,
+                            const std::string &_msgTypeName,
+                            std::vector<std::string> &_uuids)
+{
+  using HandlerTPtr = std::shared_ptr<HandlerT>;
+  std::map<std::string, std::map<std::string, HandlerTPtr>> handlers;
+
+  _handlerStorage.Handlers(_fullyQualifiedTopic, handlers);
+  for (const auto &collection : handlers)
+  {
+    for (const auto &collectionEntry : collection.second)
+    {
+      const HandlerTPtr &handler = collectionEntry.second;
+      const std::string &handlerMsgType = handler->TypeName();
+      if (handlerMsgType == _msgTypeName
+          || handlerMsgType == kGenericMessageType)
+      {
+        _uuids.push_back(handler->NodeUuid());
+      }
+    }
+  }
+}
+
+//////////////////////////////////////////////////
+std::vector<std::string> NodeShared::HandlerWrapper::NodeUuids(
+    const std::string &_fullyQualifiedTopic,
+    const std::string &_msgTypeName) const
+{
+  std::vector<std::string> uuids;
+  AppendNodeUuids(this->normal, _fullyQualifiedTopic, _msgTypeName, uuids);
+  AppendNodeUuids(this->raw, _fullyQualifiedTopic, _msgTypeName, uuids);
+
+  return uuids;
+}
+
+//////////////////////////////////////////////////
+bool NodeShared::HandlerWrapper::RemoveHandlersForNode(
+    const std::string &_fullyQualifiedTopic,
+    const std::string &_nUuid)
+{
+  bool removed = false;
+  removed |= this->normal.RemoveHandlersForNode(_fullyQualifiedTopic, _nUuid);
+  removed |= this->raw.RemoveHandlersForNode(_fullyQualifiedTopic, _nUuid);
+
+  return removed;
+}
+
+
+//////////////////////////////////////////////////
 void NodeSharedPrivate::SecurityOnNewConnection()
 {
   std::string user, pass;
@@ -1200,7 +1348,10 @@ void NodeSharedPrivate::AccessControlHandler()
         sequence = receiveHelper(*sock);
         domain = receiveHelper(*sock);
         address = receiveHelper(*sock);
+
+        // cppcheck-suppress unreadVariable
         routingId = receiveHelper(*sock);
+
         mechanism = receiveHelper(*sock);
         givenUsername = receiveHelper(*sock);
         givenPassword = receiveHelper(*sock);
@@ -1214,6 +1365,14 @@ void NodeSharedPrivate::AccessControlHandler()
         // std::cout << "Mechanism[" << mechanism << "]\n";
         // std::cout << "Username[" << givenUsername << "] [" << user << "]\n";
         // std::cout << "Pass[" << givenPassword << "] [" << pass << "]\n";
+
+        // Check that we received some kind of address. This could be used
+        // in the future to only accept connections from specific addresses.
+        if (address.empty())
+        {
+          sendAuthErrorHelper(*sock, "Invalid address");
+          continue;
+        }
 
         // Check the version
         if (version != "1.0")
@@ -1269,3 +1428,4 @@ void NodeSharedPrivate::AccessControlHandler()
 
   sock->close();
 }
+